Server MOTD added
Also added a new setting - shortname - which needs a better description.
I've based it off ETS2MP's API, which there's the server name which
would be the extended one "[SA] RockyTV's Dev Server | MODS/STUFF", and
there's the shortname, which would be "RockyTV's Test Server".

I added it in a new message because that was the best way I've found to
finally learn how to send messages through the DMP network.