Default map filter
Set the FlightState.mapViewFilterState to default at -1026.  This value
maps to all categories except debris and flags being visible.  For some
reason, this caused an occasional issue where no vessels would be
visible in the tracking station even though the filter buttons were
correctly highlighted.  The temporary work around was to fly a vessel,
or toggle any of the buttons manually, and then vessels would be visible
again.